+++ src/core/codecs/lucene99/lucene99_flat_vectors_writer.rs
@@ -245,8 +245,8 @@ where
       &mut self.vector_data,
       &field_data.field_info,
       max_doc,
-      vector_data_offset as i64,
-      vector_data_length as i64,
+      vector_data_offset,
+      vector_data_length,
       field_data.get_docs_with_field_set(),
     )?;
 
@@ -299,8 +299,8 @@ where
       &mut self.vector_data,
       &field_data.field_info,
       max_doc,
-      vector_data_offset as i64,
-      vector_data_length as i64,
+      vector_data_offset,
+      vector_data_length,
       &new_docs_with_field,
     )?;
 
@@ -389,8 +389,8 @@ where
       &mut self.vector_data,
       field_info.as_ref(),
       merge_state.segment_info.max_doc()?,
-      vector_data_offset as i64,
-      vector_data_length as i64,
+      vector_data_offset,
+      vector_data_length,
       &docs_with_field,
     )?;
     Ok(())
@@ -566,8 +566,8 @@ where
           &mut self.vector_data,
           field_info,
           merge_state.segment_info.max_doc()?,
-          vector_data_offset as i64,
-          vector_data_length as i64,
+          vector_data_offset,
+          vector_data_length,
           &docs_with_field,
         )?;
         success = true;
@@ -631,8 +631,8 @@ fn write_meta<O>(
   vector_data: &mut O,
   field: &FieldInfo,
   max_doc: i32,
-  vector_data_offset: i64,
-  vector_data_length: i64,
+  vector_data_offset: usize,
+  vector_data_length: usize,
   docs_with_field: &DocsWithFieldSet,
 ) -> Result<()>
 where
@@ -642,8 +642,8 @@ where
   meta.write_int(field.get_vector_encoding().ordinal())?;
   meta.write_int(field.get_vector_similarity_function().ordinal())?;
 
-  meta.write_vlong(vector_data_offset)?;
-  meta.write_vlong(vector_data_length)?;
+  meta.write_vlong(vector_data_offset as i64)?;
+  meta.write_vlong(vector_data_length as i64)?;
   meta.write_vint(field.get_vector_dimension())?;
 
   // write docIDs
